backend/requirements.txt
flask
opencv-python
numpy
pytesseract
flask-cors
torch
transformers
Pillow


